+++ src/common.h
@@ -131,9 +131,13 @@ v8::Local<v8::Object> FromDescribeConsumerGroupsResult(
 v8::Local<v8::Array> FromDeleteGroupsResult(
     const rd_kafka_DeleteGroups_result_t *);
 
-// ListConsumerGroupOffsets: Request
+// ListConsumerGroupOffsets: Response
 v8::Local<v8::Array> FromListConsumerGroupOffsetsResult(
     const rd_kafka_ListConsumerGroupOffsets_result_t *result);
+
+// DeleteRecords: Response
+v8::Local<v8::Array> FromDeleteRecordsResult(
+    const rd_kafka_DeleteRecords_result_t* result);
 }  // namespace Admin
 
 namespace TopicPartition {