+++ types/kafkajs.d.ts
@@ -4,7 +4,8 @@ import {
   GroupOverview,
   LibrdKafkaError,
   GroupDescriptions,
-  DeleteGroupsResult
+  DeleteGroupsResult,
+  DeleteRecordsResult
 } from './rdkafka'
 
 // Admin API related interfaces, types etc; and Error types are common, so
@@ -14,7 +15,8 @@ export {
   GroupOverview,
   LibrdKafkaError,
   GroupDescriptions,
-  DeleteGroupsResult
+  DeleteGroupsResult,
+  DeleteRecordsResult
 } from './rdkafka'
 
 export interface OauthbearerProviderResponse {
@@ -315,6 +317,8 @@ export type FetchOffsetsPartition = PartitionOffset & { metadata: string | null,
 
 export type TopicInput = string[] | { topic: string; partitions: number[] }[]
 
+export type SeekEntry = PartitionOffset
+
 export type Consumer = Client & {
   subscribe(subscription: ConsumerSubscribeTopics | ConsumerSubscribeTopic): Promise<void>
   stop(): Promise<void>
@@ -379,4 +383,8 @@ export type Admin = {
     timeout?: number,
     requireStableOffsets?: boolean }): 
     Promise<Array<{topic: string; partitions:FetchOffsetsPartition[]}>>
+  deleteTopicRecords(options: {
+    topic: string; partitions: SeekEntry[];
+    timeout?: number; operationTimeout?: number
+  }): Promise<DeleteRecordsResult[]>
 }