+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# confluent-kafka-javascript v0.5.0
+
+v0.5.0 is a limited availability feature release. It is supported for all usage.
+
+## Enhancements
+
+1. Add support for an Admin API to delete records.(#141).
+
+
 # confluent-kafka-javascript v0.4.0
 
 v0.4.0 is a limited availability feature release. It is supported for all usage.